+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * premium feature: queestion screening/reviewing/whitelisting ([PR #26](https://github.com/liveask/liveask/pull/26))
 * server side question validation (min/max length...)
+* admin view exposes mod link in event view
 
 ### Fixes
 * continous re-renders due to data change detection in `question-age-timer`